feat: add support for changing round - move projects to root dir
