+++ apps/platform/src/app/projects/[projectId]/page.tsx
