+++ apps/platform/src/app/projects/create/page.tsx
