+++ packages/kit/src/hooks/useDebounce.ts
@@ -1,0 +1,17 @@
+'use client';
+
+import { useEffect, useState } from 'react';
+
+export function useDebounce<T>(value: T, delay = 800) {
+	const [debouncedValue, setDebouncedValue] = useState<T>(value);
+
+	useEffect(() => {
+		const timeout = setTimeout(() => {
+			setDebouncedValue(value);
+		}, delay);
+
+		return () => clearTimeout(timeout);
+	}, [value, delay]);
+
+	return debouncedValue;
+}
